Revert "Suppression de la page next / Ajout homepage a la base de la route"
